Refactoring and implementation
* Added more ops
* Moved infrastructure into a utility file
